VFIDiff_normal/requirements.txt
torch==2.1.1
xformers==0.0.23
gradio==3.50
scipy==1.9.3
numpy
opencv-python
matplotlib
timm
pandas
pillow
scikit-learn
scikit-image
lpips
loguru
omegaconf
six
tqdm
albumentations
einops
imageio
loralib
